+++ openapi/src/Generate.kt
@@ -140,8 +140,10 @@ private fun KType.toJsonContent(response: Boolean = false) = mapOf(MimeTypes.jso
 
 private fun List<SecurityRequirement>.toSecurity() = map { mapOf(it.name to it.scopes.toList()) }.takeIf { it.isNotEmpty() }
 
-internal fun <T: Annotation> T.toNonEmptyValues(filter: (KProperty1<T, *>) -> Boolean = { true }): MutableMap<String, Any?> = HashMap<String, Any?>().also { map ->
-  publicProperties.filter(filter).forEach { p ->
+internal fun <T: Annotation> T.toNonEmptyValues(filter: (KProperty1<T, *>) -> Boolean = { true }): MutableMap<String, Any?> {
+  val map = HashMap<String, Any?>()
+  val props = javaClass.interfaces.first().kotlin.publicProperties.values as Iterable<KProperty1<T, *>>
+  props.filter(filter).forEach { p ->
     when(val v = p.valueOf(this)) {
       "", "default", false, 0, Int.MAX_VALUE, Int.MIN_VALUE, 0.0, Void::class.java, AccessMode.AUTO, RequiredMode.AUTO, AdditionalPropertiesValue.USE_ADDITIONAL_PROPERTIES_ANNOTATION -> null
       is Enum<*> -> v.takeIf { v.name != "DEFAULT" && v.name != "AUTO" }
@@ -149,4 +151,6 @@ internal fun <T: Annotation> T.toNonEmptyValues(filter: (KProperty1<T, *>) -> Bo
       is Array<*> -> v.map { (it as? Annotation)?.toNonEmptyValues() ?: it }.takeIf { it.isNotEmpty() }
       else -> v
     }?.let { map[p.name] = it }
-  }}
+  }
+  return map
+}